pyproject.toml
[tool.poetry]
name = "playwrightING"
version = "0.1.0"
description = "Get your ING account data"
authors = ["Pablo Cabezas <headsrooms@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
playwright = "^1.11"
12factor-configclasses = "^0.3"
python-dotenv = "^0.15"
selectolax = "^0.2"
more-itertools = "^8.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
